// model/Smjer.ts
export enum Smjer {
  ELEKTROTEHNIKA = 'ELEKTROTEHNIKA',
  GRADITELJSTVO = 'GRADITELJSTVO',
  INFORMATIKA = 'INFORMATIKA',
  MEHATRONIKA = 'MEHATRONIKA',
  RAČUNARSTVO = 'RAČUNARSTVO',
  STROJARSTVO = 'STROJARSTVO',
  DIGITALNA_EKONOMIJA = 'DIGITALNA EKONOMIJA',
  INFORMACIJSKA_SIGURNOST = 'INFORMACIJSKA SIGURNOST'
}